+++ src/plugins/extra/backgrounds/unsplash/UnsplashCredit.tsx
@@ -1,28 +1,15 @@
 import * as React from 'react';
 import { UNSPLASH_UTM } from './constants';
 import { Image } from './interfaces';
-const playIcon = require('feather-icons/dist/icons/play.svg');
-const pauseIcon = require('feather-icons/dist/icons/pause.svg');
 
 interface Props {
   image: Image;
-  paused: boolean;
-  pause: () => void;
-  play: () => void;
 }
 
 const UnsplashCredit: React.StatelessComponent<Props> = (props) => (
   <div className="credit">
     <span style={{float: 'right'}}>
-      {props.image.location_title} &emsp;
-      {props.paused
-        ? <a onClick={props.play} title="Resume new images">
-            <i dangerouslySetInnerHTML={{ __html: playIcon }} />
-          </a>
-        : <a onClick={props.pause} title="Pause on this image">
-            <i dangerouslySetInnerHTML={{ __html: pauseIcon }} />
-          </a>
-      }
+      {props.image.location_title}
     </span>
 
     <a